+++ sdk/resourcemanager/azure-resourcemanager-containerservice/src/main/java/com/azure/resourcemanager/containerservice/models/ManagedClusterAadProfile.java
@@ -28,8 +28,7 @@ public final class ManagedClusterAadProfile {
     private Boolean enableAzureRbac;
 
     /*
-     * The list of AAD group object IDs that will have admin role of the
-     * cluster.
+     * The list of AAD group object IDs that will have admin role of the cluster.
      */
     @JsonProperty(value = "adminGroupObjectIDs")
     private List<String> adminGroupObjectIDs;
@@ -53,8 +52,8 @@ public final class ManagedClusterAadProfile {
     private String serverAppSecret;
 
     /*
-     * The AAD tenant ID to use for authentication. If not specified, will use
-     * the tenant of the deployment subscription.
+     * The AAD tenant ID to use for authentication. If not specified, will use the tenant of the deployment
+     * subscription.
      */
     @JsonProperty(value = "tenantID")
     private String tenantId;