+++ sdk/resourcemanager/azure-resourcemanager-containerservice/src/main/java/com/azure/resourcemanager/containerservice/models/WorkloadRuntime.java
@@ -8,7 +8,7 @@ import com.azure.core.util.ExpandableStringEnum;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Collection;
 
-/** Defines values for WorkloadRuntime. */
+/** Determines the type of workload a node can run. */
 public final class WorkloadRuntime extends ExpandableStringEnum<WorkloadRuntime> {
     /** Static value OCIContainer for WorkloadRuntime. */
     public static final WorkloadRuntime OCICONTAINER = fromString("OCIContainer");