+++ sdk/resourcemanager/azure-resourcemanager-containerservice/CHANGELOG.md
@@ -22,6 +22,8 @@
   in `ManagedClusterSecurityProfile`.
 - Removed class `ManagedClusterSecurityProfileAzureDefender`.
 
+### Other Changes
+
 #### Dependency Updates
 
 - Updated `api-version` to `2022-06-01`.