+++ sdk/resourcemanager/azure-resourcemanager-containerservice/CHANGELOG.md
@@ -1,10 +1,18 @@
 # Release History
 
-## 2.18.0-beta.1 (Unreleased)
+## 2.18.0 (2022-08-26)
+
+### Features Added
+
+- Supported `beginCreateAgentPool` in `KubernetesCluster`.
 
 ### Other Changes
 
 - Deprecated method `KubernetesClusters.listKubernetesVersions`. Use `KubernetesClusters.listOrchestrators`.
+
+#### Dependency Updates
+
+- Updated `api-version` to `2022-07-01`.
 
 ## 2.17.0 (2022-07-25)